+++ frontend/src/components/customer/order/index.ts
@@ -1,1 +1,3 @@
-export { default as OrderCard } from "./OrderCard";
+export * from "./NoOrder";
+export * from "./OrderCard";
+export * from "./OrderStatus";
